// src/types.ts
export interface Service {
  id: string;
  title: string;
  description: string;
  icon: string;
  path: string;
}

export interface Specialty {
  id: string;
  name: string;
  description: string;
  image: string;
  icon: string;
}

export interface Solution {
  title: string;
  description: string;
}